[Ash] feat: 첫화면 UI 디테일 수정
